Throw a relevant exception on 4xx or 5xx errors
Amazon MWS uses different 4xx or 5xx errors to indicate what type of issue we
had with the API. Originally the library just threw a single exception with
no way to tell what the issue was.

This changes it to use multiple different exceptions, so that clients can
respond appropriately.

https://docs.developer.amazonservices.com/en_US/dev_guide/DG_Errors.html