Return an object from GetReportRequestStatus
This provides all information from the request as fields of the response
rather than array. Some fields are only available if the status of the report
is in progress or done.

The code does not cover all possible processing statuses, and will currently
throw an exception when it finds those.