Check for Amazon quota headers
It seems that in some instances, Amazon does not provide the x-mws-quota headers
as documented. Before trying to access the headers we now do a check to see if
they are set.

If we don't have a quota-resets-on value, we set to the current time. We might
want to revisit this at some point in the future.